requirements.txt
snowflake-snowpark-python==1.5.1
snowflake-connector-python==3.0.4
schemachange==3.5.3
